aws_mpl_rs

Module keystore

Available on crate feature ddb only.

Structs

BeaconKeyMaterials

BranchKeyMaterials

CreateKeyInput

CreateKeyOutput

Outputs for Branch Key creation.

Discovery

GetActiveBranchKeyInput

Inputs for getting a Branch Key’s ACTIVE version.

GetActiveBranchKeyOutput

Outputs for getting a Branch Key’s ACTIVE version.

GetBeaconKeyInput

Inputs for getting a Beacon Key

GetBeaconKeyOutput

Outputs for getting a Beacon Key

GetBranchKeyVersionInput

Inputs for getting a version of a Branch Key.

GetBranchKeyVersionOutput

Outputs for getting a version of a Branch Key.

GetKeyStoreInfoOutput

The configuration information for a Key Store.

KeyStoreConfig

MrDiscovery

VersionKeyInput

Inputs for versioning a Branch Key.

VersionKeyOutput

Outputs for versioning a Branch Key.

Enums

KmsConfiguration

Configures Key Store’s KMS Key ARN restrictions. Configures Key Store’s KMS Key ARN restrictions.

Traits

KeyStore

KeyStoreAdmin

Functions

make_key_store

make_key_store_admin

Type Aliases

KeyStoreAdminRef

KeyStoreRef