aws_mpl_rs

Module kms_keyring

Available on crate feature kms only.

Structs

CreateAwsKmsDiscoveryKeyringInput

Inputs for for creating a AWS KMS Discovery Keyring.

CreateAwsKmsDiscoveryMultiKeyringInput

Inputs for for creating an AWS KMS Discovery Multi-Keyring.

CreateAwsKmsEcdhKeyringInput

Inputs for creating an AWS KMS ECDH Keyring.

CreateAwsKmsHierarchicalKeyringInputddb

Inputs for creating a Hierarchical Keyring.

CreateAwsKmsKeyringInput

Inputs for for creating a AWS KMS Keyring.

CreateAwsKmsMrkDiscoveryKeyringInput

Inputs for creating a AWS KMS MRK Discovery Keyring.

CreateAwsKmsMrkDiscoveryMultiKeyringInput

Inputs for for creating a AWS KMS MRK Discovery Multi-Keyring.

CreateAwsKmsMrkKeyringInput

Inputs for for creating an AWS KMS MRK Keyring.

CreateAwsKmsMrkMultiKeyringInput

Inputs for for creating a AWS KMS MRK Multi-Keyring.

CreateAwsKmsMultiKeyringInput

Inputs for for creating a AWS KMS Multi-Keyring.

CreateAwsKmsRsaKeyringInput

Inputs for creating a AWS KMS RSA Keyring.

DiscoveryFilter

A filter which defines what AWS partition and AWS accounts a KMS Key may be in for a Keyring to be allowed to attempt to decrypt it.

GetBranchKeyIdInput

Inputs for determining the Branch Key which should be used to wrap or unwrap the data key for this encryption or decryption

GetBranchKeyIdOutput

Outputs for the Branch Key responsible for wrapping or unwrapping the data key in this encryption or decryption.

Traits

BranchKeyIdSupplier

Functions

create_aws_kms_discovery_keyring

Creates an AWS KMS Discovery Keyring, which supports unwrapping data keys wrapped by a symmetric AWS KMS Key for a single region.

create_aws_kms_discovery_multi_keyring

Creates an AWS KMS Discovery Multi-Keyring, which supports unwrapping data keys wrapped by a symmetric AWS KMS Key, for multiple regions.

create_aws_kms_ecdh_keyring

Creates an AWS KMS ECDH Keyring, which wraps and unwraps data keys by deriving a shared data key from the established shared secret between parties through the ECDH protocol.

create_aws_kms_hierarchical_keyringddb

Creates a Hierarchical Keyring, which supports wrapping and unwrapping data keys using Branch Keys persisted in DynamoDB and protected by a symmetric AWS KMS Key or AWS KMS Multi-Region Key.

create_aws_kms_keyring

Creates an AWS KMS Keyring, which wraps and unwraps data keys using single symmetric AWS KMS Key.

create_aws_kms_mrk_discovery_keyring

Creates an AWS KMS MRK Discovery Keyring, which supports unwrapping data keys wrapped by a symmetric AWS KMS Key or AWS KMS Multi-Region Key in a particular region.

create_aws_kms_mrk_discovery_multi_keyring

Creates an AWS KMS MRK Discovery Multi-Keyring that supports unwrapping data keys wrapped by a symmetric AWS KMS Key or AWS KMS Multi-Region Key, for a single region.

create_aws_kms_mrk_keyring

Creates an AWS KMS MRK Keyring, which wraps and unwraps data keys using single symmetric AWS KMS Key or AWS KMS Multi-Region Key.

create_aws_kms_mrk_multi_keyring

Creates an AWS KMS MRK Multi-Keyring, which wraps and unwraps data keys using one or more symmetric AWS KMS Keys or AWS KMS Multi-Region Keys.

create_aws_kms_multi_keyring

Creates an AWS KMS Multi-Keyring, which wraps and unwraps data keys using one or more symmetric AWS KMS Keys.

create_aws_kms_rsa_keyring

Creates an AWS KMS RSA Keyring, which wraps and unwraps data keys using a single asymmetric AWS KMS Key for RSA.

Type Aliases

AccountId

BranchKeyIdSupplierRef

GrantTokenList

KmsKeyId

Region